+++ CHANGELOG.md
@@ -7,7 +7,8 @@
 - PR #1286: Importing treelite models as FIL sparse forests
 - PR #1285: Fea minimum impurity decrease RF param
 - PR #1301: Add make_regression to generate regression datasets
-
+- PR #1289: Add Python benchmarking support for FIL
+ 
 ## Improvements
 - PR #1170: Use git to clone subprojects instead of git submodules
 - PR #1239: Updated the treelite version
@@ -22,6 +23,7 @@
 - PR #1300: Lowercase parameter versions for FIL algorithms
 - PR #1312: Update CuPy to version 6.5 and use conda-forge channel
 - PR #1314: Added options needed for ASVDb output (CUDA ver, etc.), added option to select algos
+- PR #1340: Updated unit tests to uses larger dataset
 
 ## Bug Fixes
 - PR #1212: Fix cmake git cloning always running configure in subprojects